src/car-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
httpx
pydantic
python-dateutil
pytest
pytest-asyncio 
../auth-service
